+++ alicompat/fabric/src/compat/repurposedstructures/java/com/yanny/alicompat/compat/repurposedstructures/RepurposedStructuresCompat.java
@@ -1,9 +1,6 @@
 package com.yanny.alicompat.compat.repurposedstructures;
 
-import com.telepathicgrunt.repurposedstructures.misc.maptrades.StructureSpecificMaps;
-import com.yanny.ali.api.IServerRegistry;
 import com.yanny.alicompat.IModCompat;
-import com.yanny.alicompat.accessor.PluginUtils;
 import org.jetbrains.annotations.NotNull;
 
 public class RepurposedStructuresCompat implements IModCompat {
@@ -13,10 +10,5 @@ public class RepurposedStructuresCompat implements IModCompat {
     @Override
     public String targetModId() {
         return MOD_ID;
-    }
-
-    @Override
-    public void registerServer(IServerRegistry registry) {
-        PluginUtils.registerItemListing(registry, StructureSpecificMaps.TreasureMapForEmeralds.class, TreasureMapForEmeraldsAccessor.class);
     }
 }