+++ alicompat/CHANGELOG.md
@@ -5,11 +5,14 @@
 - Added Artifacts support
 - Added Dimensional Dungeons support
 - Added ExtraStorage support
+- Added Grimoire of Gaia support
 - Added Moonlight support
 - Added More Red support
 - Added Occultism support
 - Added Placebo support
 - Added Refined Storage support
+- Added Repurposed Structures support
+- Added Sawmill support
 - Added Snow! Real Magic! support
 - Added Sophisticated Backpacks support
 - Added Sophisticated Storage support